Revert "새로운 숙소 생성 기능 구현"
