--- airdnb/src/main/java/com/example/airdnb/dto/accommodation/AddressRequest.java
@@ -1,4 +1,0 @@
-package com.example.airdnb.dto.accommodation;
-
-public record AddressRequest(String country, String state, String city, String detail, Double latitude, Double longitude) {
-}
